BUG: Propagate the `target` object from the root in extended trait change notifications.
The `TraitChangeNotificationWrapper`s created for the subsequent segments in an
extended trait change notification do not know when they need to clean
themselves up. Sometimes, this is taken care of when the `owner` and the
`target` of the root `ListenerNotifyWrapper` are different objects. `Property`s
with `depends_on`, on the other hand` have both objects the same, so they never
get a chance to unregister.