Fixed th Array traits's validator
Fixed the Array trait's validator which incorrectly required the value
to be either a ndarray instance or a SequenceType. This is inconsistent
with the documentation, where it says that Array traits behave exactly
as if the argument was given to np.array, which however accepts a lot
more than just sequence types. In particular, scalar numbers should
be considered a zero-dimensional array instead of beeing rejected.
An array scalar is perfectly fine as ndarray.