Cover `ExtendedTraitChangeNotifyWrapper`.
Much of the code is duplicated from `test_dynamic_notifiers.py`. However,
I considered having a single class and dynamically setting the `dispatch`
argument to `@on_trait_change`, and found that it appeared as if
`dispatch="extended"` was a legitimate setting of the dispatch parameter,
while in fact I was just relying on some implementation details.

I find that, in this case, the duplication makes the tests self-contained
and more explicit.