The `dispatch='ui'` option has been broken since 2007, remove it.
Since 2007, the documentation has given the illusion that it would be
possible to queue a trait notification on the UI thread. In reality,
the functionality was broken and the notification always occurred
immediately on the UI thread. The "fast_ui" and "ui" traits dispatch
mechanism are thus equivalent. Remove the code that is not used.