Add extra `gc.collect()`.
Without this extra call to `gc.collect`, running the test fails sometimes
depending on the order in which the tests are called. The failure says
that there is one object *less* at the end of the test than at the
beginning. Clearly, there are object that had not been collected.