MAINT: Re-raise AttributeError on the Cython side.
When the call to PyObject_GenericGetAttr fails for an attribute lookup,
NULL is returned and PyErr is set. When getattr_python finishes, None is
returned and Python terminates, since attribute lookup resulted in an
value (None) but with the error flag set.

Perhaps there's a better way of "raising" the exception on the Python
side, but the only way I could get this to work was by clearing the
error, and manually re-raising it in Cython.