CLN: Add import check helper (#521)
* Add function to handle import errors

* Add test file and license header

* Refactor numpy import checks

* Remove boilerplate, and some unused imports

* Add test for import handler

* Update docstring

* Add None test

* Add space to header

* Fix mistake in return value

* Add missing imports

* Fix mistake in renaming function

* Fake to trigger build

* Remove addition that triggered build

* Remove sphinx import boilerplate

* Move import_handler module to testing.

We shouldn't make import_handler a publicly available module just
yet; we don't want to encourage other packages to import from it.
Since it's only used in testing, let's move this utility to
the testing package.

* Simplify signature of import_handler

The value returned by import_handler is redundant, since the boolean
part can be inferred from the module part. We simplify the return
value to be only the module or `None`.

* Fix: module name is not necessarily lower case.

* Move the requires_* decorators to a common location.

* Rename import_handler module.

Rename the module and corresponding test module; the
contents of the module have changed significantly. At
this point, that module is essentially capturing the
optional dependencies used during testing.

* Rename function to optional_import

Rename "import_handler" to "optional_import", to give
the user a hint about what it does.

* Add requires_python2 and conflicts_with_python2 decorators.

* Remove the mutation of global state.

* Restore unrelated changes.

* Restore one other unrelated change.