Fix corner-case behaviour for __iadd__ and extend
For __iadd__ and extend, we need to make sure that we validate
_all_ incoming items before we append any of them, else we
risk both modifying the list _and_ raising. So replace
the corresponding map() calls with list(map()), and convert
to a list comprehension for readability.

Also converts one other occurrence of list(map()) to a list
comprehension, with no change in semantics.