FIX: Dont configure the sphinx copybutton extension in docs
This PR simply removes the sphinx-copybutton extension-specific
configuration from the sphinx configuration file.

this is because the configuration was in some cases returning bad data.
the prompt text configuration seems complicated and difficult to get
right to support the wide range of code samples we have in the
documentation.

for now, having a copy button that copies the entire code sample is
better than a smart copy button that only copies the "right" piece of
code

	modified:   docs/source/conf.py