Improve Int and Float type stubs (#1656)
This PR improves the type stubs for Int and Float, being more accurate about what's accepted for those trait types.

We need SupportsIndex, which is only available in typing from Python 3.8 onwards. So I've added typing-extensions as a dependency for traits-stubs. We can drop the dependency once we no longer support Python 3.6 and Python 3.7.