Add pyproject.toml (#1676)
Long overdue PR to add a pyproject.toml, containing build configuration and basic black and isort configurations, for those who want to use black and isort on new files in this repository. (Note that many existing files don't conform to those configurations. We have no immediate plans to rectify that, but the configurations may be useful for new code.)

With the build dependencies explicitly mentioned in pyproject.toml, there should be no need to pre-install setuptools and/or wheel in the test environments, so we've updated the workflows accordingly.