Remove outdated alias
This PR removes an `enumerate = enumerate` assignment that should no longer be necessary.

Technically, this is a backwards incompatible change: anyone who imports `enumerate` from `traits.trait_base` will no longer be able to do so. However, the workaround is immediate (just delete that import so that you pick up the built-in `enumerate`), and code that's doing this is already wrong on so many levels, so I don't feel bad about making this change without a deprecation period.