Use PySide6 instead of PySide2 in CI testing (#1685)
This PR changes the .[test] setup.py requirements to use PySide6 instead of PySide2 for Python3.8 and later, and removes the version constraint on the PySide install. We exclude Python 3.11, since the PySide6 eggs available on PyPI claim not to work on 3.11.

While we're here, we also clean up an outdated Sphinx version restriction.