Remove deleter for the kiva_backend property (which is readonly) (#1686)
PR #1670 added deleters for various ETSConfig attributes. However, the kiva_backend property is readonly, with value derived from the toolkit attribute, so we shouldn't be providing a deleter for it. The right way to mock kiva_backend is to set the toolkit attribute (and in fact there's already a test for that).

This PR removes the deleter for the kiva_backend attribute, and makes the `kiva_backend` property getter a pure function with no caching.

(cherry picked from commit aa05dc6f5a4d920e4a3b184ca727999704042afc)