Don't collect traits-stubs tests under pytest (#1690)
This PR adds a pytest configuration to ignore traits-stubs when running the Traits tests. (It's fine to run the traits-stubs tests under pytest, but it needs to be done from the traits-stubs directory as a separate test run.)

Together with #1684 and #1683, this closes #1288.

(cherry picked from commit 18341f74aa741c15160ea934a994f70cf6bbd870)