Make ETSConfig class documentation visible in the API docs (#1688)
This PR provides an alternative to #1669. The goal is to make the ETSConfig docstrings available in the API docs, so that a search for "ETSConfig" in the online documentation gives useful results.

To that end, this PR:

* Renames the ETSConfig class to ETSConfigType
* Keeps the ETSConfig instance the same as before (previously, the created instance shadowed the class)
* Updates the main docstrings of ETSConfigType and ETSConfig.

There shouldn't be backwards-compatibility concerns with the rename: the ETSConfig class wasn't available before (since it was shadowed by the instance), and the trick of using type(ETSConfig) to get the class will still work.

I'm deliberately not exposing ETSConfigType in traits.api, since in most circumstances it shouldn't be used directly

(cherry picked from commit 9e4664ccd452bef3ec4692759f75a2b698c86ca8)