Fix TraitType clone default values (#1645)
This PR works towards a fix of #1630 and related issues.

The starting point is a set of regression tests that identify the behaviour which doesn't seem correct and test for the expected behaviour. This then adds changes to TraitTypes.clone that fix a number of the observed issues. We then remove the default value wrangling in HasTraits as it should no longer be needed.

I think at this point everything is in a better state than it was: we have less unexpected behaviour.

There may be subsequent things which might be advisable, such as raising warnings about future deprecation of certain behaviours, adding clone methods to particular trait types, etc.

Co-authored-by: Mark Dickinson <mdickinson@enthought.com>
(cherry picked from commit 7ecd065637fa935f9f669dc9b2cd7628f2cbcce3)