Fix and rename requires_numpy_typing decorator (#1710)
PR #1682 introduced what should have been a check for numpy.typing, that ended up as a check for numpy.testing instead. This PR fixes that.

Also, PR #1682 should have failed on Python 3.6 because of the above, since on Python 3.6 the latest version of NumPy available doesn't have type hints. It didn't fail because we weren't populating the package data with the new numpy_examples directory, so the numpy-based tests weren't run. That's fixed in #1709.