MAINT: add main thread check for ui dispatch, solve no ui failure (#1740)
Previously the ui_dispatch(handler, *args, **kw) function in
traits/traits/trait_notifiers.py only checks current_thread().ident with
ui_thread. But supposedly, this current thread id should be compared
with the main thread (threading.main_thread()). This discrepancy caused
a no ui failure in ui dispatch as mentioned in issue #1732. This PR
attempts to resolve this failure. Closes #1732

**Checklist**
- [ ] Tests
- [x] Update API reference (`docs/source/traits_api_reference`) (seems
no need to change for this PR)
- [ ] Update User manual (`docs/source/traits_user_manual`) (seems no
need to change for this PR)
- [x] Update type annotation hints in stub files (seems no need to
change for this PR)

---------

Co-authored-by: Chengyu Liu <cyliu@aus552cyliu.local>
Co-authored-by: Mark Dickinson <mdickinson@enthought.com>