Fix some compiler warnings (#1774)
This PR fixes some compiler warnings that I noticed in the failed
Windows build in PR #1773. There are two classes of fixes:

- Fix occurrences of `int` used to represent the size of a tuple / list
/ dict. Those should be `Py_ssize_t`
- Fix two functions that treat a mask as having type `int`, when it
should be `unsigned int`. All the actual masks used have type `unsigned
int`.

The warnings themselves are relatively harmless, but silencing them
gives us a better chance of noticing warnings that we _should_ be taking
notice of.

The PR also adds a `-v` flag to the `pip install` commands that install
Traits, so that compiler warnings will always be visible in the build
log. Without the `-v`, we only get to see the warnings for failed
builds.