Fix warning messages from PyDict_GetItem (#1818)
This PR silences warning messages from `PyDict_GetItem` in cases where
we're trying to look up something that's not hashable.

Note that replacing `PyDict_GetItem` with `PyDict_GetItemWithError`
means replacing something that can't raise an exception with something
that can, but in both cases we already have error paths present that
handle the case where an exception is set. (In the `validate_trait_map`
case, `raise_traits_error` takes care of clearing any existing error.
Arguably it should chain the errors instead, but that's a separate
discussion.)

Closes #1817