Fix the wheel-building workflow (#1822)
This PR attempts to fix the current wheel-building workflow. That
workflow is failing on manylinux-aarch64 builds with a gcc internal
compiler error, so until we can understand the problem better we'll drop
those builds and the QEMU emulation.

While we're here, we also perform separate macOS Intel and ARM builds.