Fix that clone_traits doesn't respect dynamic traits (#1846)
Closes #1845

This PR fixes `clone_traits` (and `copy_traits`) to respect dynamic
traits.

Previously, the `HasTraits.copy_traits` method used a mixture of trait
information from `self` (the object being copied _to_) and `other` (the
object being copied _from_) to determine how to behave. Since the actual
copying was enclosed in a large `try / except` block, any errors arising
just resulted in the trait value not being copied.

In particular, when cloning a `HasTraits` object with dynamically-added
traits:

- the set of traits to copy was first determined by inspecting the
object, and included those dynamically-added traits
- then, in `copy_traits`, we inspect the cloned object that we're
copying _to_, which doesn't have the dynamically-added traits, and as a
result we fail to copy those trait values.

This PR updates the logic to only use the trait information from `other`
when deciding what / how to copy, instead of using a mixture of sources.
As a result, `clone_traits` should behave as expected on objects with
dynamically-added traits.

We have a large corpus of existing Traits code, and like all changes to
Traits this change risks unintended consequences and surprise breakages
of existing code that's inadvertently relying on the existing behaviour.
While I think it's the right change to make, I'd advocate not
backporting this change to any bugfix branch, and highlighting the
change prominently in release notes.