Make Instance type hints more useful (#1840)
Closes #1839 
Closes #1764 
Closes #1673 

This PR provides a better type-hinting experience for `Instance` (and
`BaseInstance`) traits:

- fixes inclusion of `str` in the inferred type for the value of an
`Instance` trait
- adds an overload for the case `allow_none=False`, so that type
checkers can infer that the value of the trait is not `None` (and `None`
may not be assigned) in this case. Note: this technically isn't true
(since a default of `None` can still be provided even when
`allow_none=False`), but it's a safe bet that if the user is explicitly
stating `allow_none=False` then the intent is that the trait value is
never `False`
- adds an overload for the forward reference case where the class is
given as a string - we just punt in this case and report `Any` for the
value trait (for both getting and setting).